+++ lib/msal-common/test/utils/ProtocolUtils.spec.ts
@@ -1,5 +1,5 @@
 import { ProtocolUtils } from "../../src/utils/ProtocolUtils";
-import { DECRYPTED_BOUND_RT_AUTHENTICATION_RESULT_DEFAULT_SCOPES, RANDOM_TEST_GUID, TEST_CONFIG, TEST_POP_VALUES } from "../test_kit/StringConstants";
+import { DECRYPTED_BOUND_RT_AUTHENTICATION_RESULT_DEFAULT_SCOPES, RANDOM_TEST_GUID, TEST_CONFIG, TEST_CRYPTO_VALUES, TEST_POP_VALUES } from "../test_kit/StringConstants";
 import { ICrypto, PkceCodes } from "../../src/crypto/ICrypto";
 import { Constants } from "../../src/utils/Constants";
 import sinon from "sinon";
@@ -56,6 +56,9 @@ describe("ProtocolUtils.ts Class Unit Tests", () => {
             },
             async clearKeystore(): Promise<boolean> {
                 return Promise.resolve(true);
+            },
+            async hashString(): Promise<string> {
+                return Promise.resolve(TEST_CRYPTO_VALUES.TEST_SHA256_HASH);
             },
             async getAsymmetricPublicKey(): Promise<string> {
                 return TEST_POP_VALUES.DECODED_STK_JWK_THUMBPRINT;