+++ lib/msal-node/test/utils/TestConstants.ts
@@ -84,6 +84,10 @@ export const DEFAULT_CRYPTO_IMPLEMENTATION: ICrypto = {
         const notImplErr = "Crypto interface - signJwt() has not been implemented";
         throw AuthError.createUnexpectedError(notImplErr);
     },
+    async hashString(): Promise<string> {
+        const notImplErr = "Crypto interface - hashString() has not been implemented";
+        throw AuthError.createUnexpectedError(notImplErr);
+    },
     async getAsymmetricPublicKey(): Promise<string> {
         const notImplErr = "Crypto interface - getAsymmetricPublicKey() has not been implemented";
         throw AuthError.createUnexpectedError(notImplErr);