+++ lib/msal-angular/CHANGELOG.md
@@ -1,8 +1,25 @@
 # Change Log - @azure/msal-angular
 
-This log was last generated on Mon, 01 Nov 2021 23:53:22 GMT and should not be manually modified.
+This log was last generated on Tue, 04 Jan 2022 00:20:29 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 2.1.0
+
+Tue, 04 Jan 2022 00:20:29 GMT
+
+### Minor changes
+
+- Bump @azure/msal-browser to v2.21.0
+
+## 2.0.6
+
+Tue, 07 Dec 2021 00:17:01 GMT
+
+### Patches
+
+- Fix angular guard /code check to ensure proper length #4249 (janutter@microsoft.com)
+- Bump @azure/msal-browser to v2.20.0
 
 ## 2.0.5